+++ src/components/Recipe/RecipeSummary.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { StyleProp, ViewStyle, StyleSheet, Text, View } from 'react-native';
 
 import { Colors, Typography } from 'styles';
-import { TouchableOpacity } from 'components/common/Button';
+import { TouchableOpacity } from 'components/common/Base';
 import FavoriteButton from './FavoriteButton';
 import RecipeTag from 'components/Recipe/RecipeTag';
