+++ src/components/common/Base/index.ts
@@ -1,0 +1,4 @@
+import TouchableOpacity from './TouchableOpacity';
+import TouchableWithoutFeedback from './TouchableWithoutFeedback';
+
+export { TouchableOpacity, TouchableWithoutFeedback };
