+++ src/components/common/Button/index.ts
@@ -1,15 +1,6 @@
-import TouchableOpacity from './TouchableOpacity';
-import TouchableWithoutFeedback from './TouchableWithoutFeedback';
 import AddButton from './AddButton';
 import RemoveButton from './RemoveButton';
 import SingleButton from './SingleButton';
 import DoubleButton from './DoubleButton';
 
-export {
-  TouchableOpacity,
-  TouchableWithoutFeedback,
-  AddButton,
-  RemoveButton,
-  SingleButton,
-  DoubleButton,
-};
+export { AddButton, RemoveButton, SingleButton, DoubleButton };
