+++ src/screens/Home/HomeMenu.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Text, StyleSheet, StyleProp, ViewStyle } from 'react-native';
 import { SvgProps } from 'react-native-svg';
 
-import { TouchableOpacity } from 'components/common/Button';
+import { TouchableOpacity } from 'components/common/Base';
 import { Colors, Dimens, Typography } from 'styles';
 import { IS_SCREEN_HEIGHT_LESS_THAN_700 } from 'utils/PlatformInfo';
